requirement.txt
flask==3.0.0
gunicorn==21.2.0
google-ads==23.1.0
google-cloud-secret-manager==2.16.3
openai==1.13.0
python-dotenv==1.0.0
requests==2.31.0
